+++ src/features/reader/hooks/useReaderSetChaptersState.ts
@@ -11,11 +11,7 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import { Chapters } from '@/features/chapter/services/Chapters.ts';
 import type { requestManager } from '@/lib/requests/RequestManager.ts';
 import type { GetChaptersReaderQuery } from '@/lib/graphql/generated/graphql.ts';
-import type {
-    IReaderSettings,
-    ReaderOpenChapterLocationState,
-    ReaderStateChapters,
-} from '@/features/reader/Reader.types.ts';
+import type { IReaderSettings, RouteStateReader, ReaderStateChapters } from '@/features/reader/Reader.types.ts';
 import { filterChapters } from '@/features/chapter/utils/ChapterList.util.tsx';
 import type { ChapterListFilterOptions } from '@/features/chapter/Chapter.types.ts';
 import { getReaderChapterFromCache } from '@/features/reader/Reader.utils.ts';
@@ -24,6 +20,7 @@ import { getReaderChaptersStore } from '@/features/reader/stores/ReaderStore.ts'
 import { STABLE_EMPTY_OBJECT } from '@/base/Base.constants.ts';
 
 import { READER_DEFAULT_CHAPTERS_STATE } from '@/features/reader/stores/ReaderChaptersStore.ts';
+import { AppRoutes } from '@/base/AppRoute.constants.ts';
 
 export const useReaderSetChaptersState = (
     chaptersResponse: ReturnType<typeof requestManager.useGetMangaChapters<GetChaptersReaderQuery>>,
@@ -36,7 +33,7 @@ export const useReaderSetChaptersState = (
     chapterListOptions: ChapterListFilterOptions,
 ) => {
     const navigate = useNavigate();
-    const locationState = useLocation<ReaderOpenChapterLocationState>().state;
+    const locationState = useLocation<RouteStateReader>().state;
 
     const { updateInitialChapter } = locationState ?? STABLE_EMPTY_OBJECT;
     const finalInitialChapter = updateInitialChapter ? undefined : initialChapter;
@@ -77,7 +74,10 @@ export const useReaderSetChaptersState = (
         const hasInitialChapterChanged = newInitialChapter != null && newInitialChapter.id !== finalInitialChapter?.id;
 
         if (hasInitialChapterChanged) {
-            navigate('', { replace: true, state: { ...locationState, updateInitialChapter: undefined } });
+            navigate('', {
+                replace: true,
+                state: { ...locationState, ...AppRoutes.reader.state({ updateInitialChapter: undefined }) },
+            });
         }
 
         getReaderChaptersStore().setReaderStateChapters((prevState) => {